requirements.txt
streamlit
langchain-core
langchain
langchain-community
langchain-google-genai
chromadb
pypdf
pymupdf
python-dotenv